Initial Design
For our Solar cooler, we started first step to design it in the Solid Works, we are using the perfect dimension of Tesla model 3 cavity. The dimesions are as follow, Cavity Top: Length= 27 Inches, Width= 18 Inches, Cavity Bottom: Length= 24 Inches, Width= 15 Inches, Total Height= 12 Inches. For these dimensions, we draw a basic design idea of a cooler box. The main idea about this deigns to bring our thoughts on to some real model so we chose the solid works to implement it. All these files contain the basic design of our solar cooler.