Assembly 01
In this design, we improve over design and try to make it more perfect for the setup of electronics devices. In this, we make a separate part for the cooling section and separate part for electronic section. In the middle area, we are going to use as storage of food and beverages. Both sides left and the right side we are going to design a removable portion which can contain all the electronics assembly. The main cooling box area is Length = 13 Inches , Width = 12 inches, Height= 12 Inches. The two separate portion for electronics components each box has a size of Top: Length= 05 Inches, Width= 16 Inches, Cavity Bottom: Length= 04 Inches, Width= 13 Inches, Total Height= 12 Inches.